+++ src/init.c
@@ -27,11 +27,12 @@ SEXP wk_c_problems_handler_new(void);
 SEXP wk_c_sfc_writer_new(void);
 SEXP wk_c_trans_affine_new(SEXP trans_matrix);
 SEXP wk_c_trans_affine_as_matrix(SEXP trans_xptr);
+SEXP wk_c_trans_explicit_new(SEXP xy, SEXP use_z, SEXP use_m);
 SEXP wk_c_trans_set_new(SEXP xy, SEXP use_z, SEXP use_m);
 SEXP wk_c_trans_filter_new(SEXP handler_xptr, SEXP trans_xptr);
 SEXP wk_c_wkb_is_na(SEXP geom);
 SEXP wk_c_wkb_is_raw_or_null(SEXP geom);
-SEXP wk_c_wkb_to_hex(SEXP geom);
+SEXP wk_c_wkb_to_hex(const SEXP geom);
 SEXP wk_c_vertex_filter_new(SEXP handler_xptr, SEXP add_details);
 SEXP wk_c_handler_void_new(void);
 SEXP wk_c_handler_addr(SEXP xptr);
@@ -63,6 +64,7 @@ static const R_CallMethodDef CallEntries[] = {
   {"wk_c_sfc_writer_new", (DL_FUNC) &wk_c_sfc_writer_new, 0},
   {"wk_c_trans_affine_new", (DL_FUNC) &wk_c_trans_affine_new, 1},
   {"wk_c_trans_affine_as_matrix", (DL_FUNC) &wk_c_trans_affine_as_matrix, 1},
+  {"wk_c_trans_explicit_new", (DL_FUNC) &wk_c_trans_explicit_new, 3},
   {"wk_c_trans_set_new", (DL_FUNC) &wk_c_trans_set_new, 3},
   {"wk_c_trans_filter_new", (DL_FUNC) &wk_c_trans_filter_new, 2},
   {"wk_c_wkb_is_na", (DL_FUNC) &wk_c_wkb_is_na, 1},